pyproject.toml
[build-system]
requires = ["setuptools>=61, a a a"]
build-backend = "setuptools.build_meta"

[project]
name = "debtshield_ai"
version = "1.0.0"
description = "Adversarial Multi-Agent Debt Negotiation Environment"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai",
    "pydantic",
    "fastapi",
    "uvicorn",
    "python-dotenv",
    "pymupdf",
    "pyyaml",
    "openenv-core>=0.2.0",
]

[project.scripts]
server = "server.main:main"